projects/LANS-scubelic-asucur-lborovic-nmarjanovic/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-answer-watcher"
version = "0.2.0"
description = "Monitor how LLMs talk about your brand vs competitors in buyer-intent queries"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [
    { name = "Nikola Balić", email = "niko@example.com" }
]
keywords = ["llm", "monitoring", "brand", "seo", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "tenacity>=8.0",
    "rapidfuzz>=3.0",
    "jinja2>=3.1",
    "typer>=0.12.0",
    "rich>=13.0",
    "steel-sdk>=0.1.0",
    "playwright>=1.40.0",
    "mkdocs-llmstxt>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14.0",
    "freezegun>=1.5.0",
    "ruff>=0.4.0",
]

[project.scripts]
llm-answer-watcher = "llm_answer_watcher.cli:app"

[project.urls]
Homepage = "https://github.com/nibzard/llm-answer-watcher"
Documentation = "https://github.com/nibzard/llm-answer-watcher#readme"
Repository = "https://github.com/nibzard/llm-answer-watcher"
Issues = "https://github.com/nibzard/llm-answer-watcher/issues"

[tool.hatch.build.targets.wheel]
packages = ["llm_answer_watcher"]
artifacts = ["llm_answer_watcher/py.typed"]

[tool.hatch.build.targets.sdist]
include = [
    "llm_answer_watcher/",
    "tests/",
    "examples/",
    "*.md",
    "LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["llm_answer_watcher"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.35.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.3",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocs-material-extensions>=1.3.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.2",
]
